Move rocket.js into weapons/ directory
